# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-commerce Web API",
    "version": "1.0.0",
    "description": "\n## E-commerce Web Platform API Documentation\n\nThis API provides comprehensive endpoints for managing an e-commerce platform including:\n- **Authentication & Authorization** - JWT-based authentication with role-based access control\n- **User Management** - Complete user CRUD operations with admin privileges\n- **Admin Functions** - Administrative tools and management features\n- **Notifications** - Real-time notification system\n- **Health Checks** - System monitoring and health endpoints\n\n### Authentication\nMost endpoints require authentication via JWT token. Include the token in the Authorization header:\n`Authorization: Bearer your-jwt-token`\n\n### Rate Limiting\nSome endpoints (like login) have rate limiting applied to prevent abuse.\n\n### Roles\n- **USER**: Regular user with basic permissions\n- **ADMIN**: Administrative user with elevated permissions\n- **SUPER_ADMIN**: Super administrator with full system access\n      ",
    "contact": {
      "name": "E-commerce Web API Support",
      "email": "support@ecommerce-web.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer {token}"
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token",
        "description": "Authentication via HTTP-only cookie"
      }
    },
...
  },
...
  "security": [
    {
      "BearerAuth": []
    },
    {
      "CookieAuth": []
    }
  ],
  "paths": {
...
    "/api/auth/sso_google": {
      "post": {
        "summary": "Begin Google OAuth (returns redirect URL)",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "locale": {
                    "type": "string",
                    "enum": [
                      "vi",
                      "en"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Redirect URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "redirectUrl": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Google OAuth callback (redirects to frontend)",
        "tags": [
          "Authentication"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "schema": {
              "type": "string",
              "description": "locale"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to frontend with cookies"
          }
        }
      }
    },
...
  }
}